Move nu_protocol::process::ExitStatus to nu_system::ExitStatus
This will allow `nu_system::ForegroundChild::wait()` to return our
custom `ExitStatus` instead of `std::process::ExitStatus`.

Since `ExitStatus::check_ok()` depends on `Span` and `ShellError`, both
of which are defined in `nu_protocol`, there are two options:

1. Also move `Span` and `ShellError` to `nu_system`.
2. Refactor `ExitStatus` to keep `check_ok()` in `nu_protocol`.

Both `Span` and `ShellError` are very central pieces, and the latter is
also part of the larger error handling infrastructure in
`nu_protocol::errors`. On the other hand, `check_ok()` is only used in
`nu_protocol::process::child`. Therefore, option 2 is the cleaner choice
here.
