+++ crates/nu-protocol/src/process/mod.rs
@@ -1,6 +1,4 @@
 //! Handling of external subprocesses
 mod child;
-mod exit_status;
 
 pub use child::*;
-pub use exit_status::ExitStatus;
